Move the case conversion logic and the `Identifier` type into a dedicated module
